+++ src/main/java/com/github/nalamodikk/register/ModNetworkingServer.java
@@ -5,12 +5,11 @@ import com.github.nalamodikk.common.network.packet.server.manatool.*;
 import com.github.nalamodikk.narasystem.nara.network.client.NaraSystemIntroMessagePacket;
 import com.github.nalamodikk.narasystem.nara.network.client.OpenNaraInitScreenPacket;
 import com.github.nalamodikk.narasystem.nara.network.server.NaraSyncPacket;
-import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.network.event.RegisterPayloadHandlersEvent;
 
-@EventBusSubscriber(modid = KoniavacraftMod.MOD_ID ,value = Dist.DEDICATED_SERVER)
+@EventBusSubscriber(modid = KoniavacraftMod.MOD_ID)
 public class ModNetworkingServer {
     public static final String VERSION = "1";